Swap Lazy for Try in spider (#21)
When deploying an app on fly.io, I noticed what seemed like a deadlock
when deploying on 1 cpu machine. No issue when scaling to 2 cpu. Not
sure if it's related, but Lazy can/will use a lot of virtual threads, so
swap those parts of spider out for plain Trys to see if it helps.